requirements.txt
ansimarkup==1.5.0
anyio==3.6.2
APScheduler==3.10.1
certifi==2022.12.7
colorama==0.4.6
dnspython==2.3.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
pip==21.3.1
pyfiglet==0.8.post1
pymongo==4.3.3
python-telegram-bot==20.1
pytz-deprecation-shim==0.1.0.post0
pytz==2022.7.1
PyYAML==6.0
rfc3986==1.5.0
setuptools==60.2.0
six==1.16.0
sniffio==1.3.0
tzdata==2022.7
tzlocal==4.3
wheel==0.37.1
